// src/api/model/apiBeycoCoffeeGrade.ts
/*
 * Copyright(c) 2009 - 2019 Abelium d.o.o.
 * Kajuhova 90, 1000 Ljubljana, Slovenia
 * All rights reserved
 * Copyright(c) 1995 - 2018 T-Systems Multimedia Solutions GmbH
 * Riesaer Str. 5, 01129 Dresden
 * All rights reserved.
 *
 * INATrace Services API
 * Abelium INATrace Services API swagger documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the openapi-typescript-angular-generator.
 * https://github.com/alenabelium/openapi-typescript-angular-generator
 * Do not edit the class manually.
 */





export interface ApiBeycoCoffeeGrade { 
    /**
     * Grade of coffee beans
     */
    type?: ApiBeycoCoffeeGrade.TypeEnum;
}

/**
 * Namespace for property- and property-value-enumerations of ApiBeycoCoffeeGrade.
 */
export namespace ApiBeycoCoffeeGrade {
    /**
     * All properties of ApiBeycoCoffeeGrade.
     */
    export enum Properties {
        /**
         * Grade of coffee beans
         */
        type = 'type'
    }

    /**
     * All possible values of type.
     */
    export enum TypeEnum {
        Na = 'Na',
        Grade1 = 'Grade1',
        Grade2 = 'Grade2',
        Grade3 = 'Grade3',
        Grade4 = 'Grade4',
        Grade5 = 'Grade5',
        Aa = 'Aa',
        A = 'A',
        Ab = 'Ab',
        B = 'B',
        C = 'C',
        Pb = 'Pb',
        X = 'X',
        E = 'E',
        Psc = 'Psc',
        Y1 = 'Y1',
        Y2 = 'Y2',
        T = 'T',
        HgHb = 'HgHb',
        ShgShb = 'ShgShb',
        Other = 'Other'
    }


    export function formMetadata() {
        return  {
            metadata: formMetadata,
            classname: 'ApiBeycoCoffeeGrade',
            vars: [
                {
                    isReadOnly: false,
                    isEnum: true,
                    datatypeWithEnum: 'ApiBeycoCoffeeGrade.TypeEnum',
                    required: false,
                    name: 'type',
                    classname: 'ApiBeycoCoffeeGrade',
                    dataType: 'string',
                    isPrimitiveType: true,
                    isListContainer: false,
                    complexType: ''
                },
            ],
            validators: {
                type: [
                ],
            }
        }
    }

  // export const ApiBeycoCoffeeGradeValidationScheme = {
  //     validators: [],
  //     fields: {
  //               type: {
  //                   validators: []
  //               },
  //     }
  // } as SimpleValidationScheme<ApiBeycoCoffeeGrade>;


}